Refactoring : Moved XmlTextToRichText conversion to separate converter
