+++ bundle/Command/ConvertXmlTextToRichTextCommand.php
@@ -5,8 +5,6 @@
 namespace EzSystems\EzPlatformXmlTextFieldTypeBundle\Command;
 
 use DOMDocument;
-use DOMXPath;
-use eZ\Publish\Core\FieldType\RichText\Converter;
 use eZ\Publish\Core\Persistence\Database\DatabaseHandler;
 use PDO;
 use Psr\Log\LoggerInterface;
@@ -14,13 +12,8 @@ use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use eZ\Publish\Core\FieldType\RichText\Converter\Aggregate;
-use eZ\Publish\Core\FieldType\XmlText\Converter\Expanding;
-use eZ\Publish\Core\FieldType\RichText\Converter\Ezxml\ToRichTextPreNormalize;
-use eZ\Publish\Core\FieldType\XmlText\Converter\EmbedLinking;
-use eZ\Publish\Core\FieldType\RichText\Converter\Xslt;
-use eZ\Publish\Core\FieldType\RichText\Validator;
 use eZ\Publish\Core\FieldType\XmlText\Value;
+use eZ\Publish\Core\FieldType\XmlText\Converter\RichText as RichTextConverter;
 
 class ConvertXmlTextToRichTextCommand extends ContainerAwareCommand
 {
@@ -29,16 +22,6 @@ class ConvertXmlTextToRichTextCommand extends ContainerAwareCommand
      */
     private $db;
     
-    /**
-     * @var \eZ\Publish\Core\FieldType\RichText\Converter
-     */
-    private $converter;
-    
-    /**
-     * @var \eZ\Publish\Core\FieldType\RichText\Validator
-     */
-    private $validator;
-
     /**
      * @var \Psr\Log\LoggerInterface
      */
@@ -50,28 +33,6 @@ class ConvertXmlTextToRichTextCommand extends ContainerAwareCommand
 
         $this->db = $db;
         $this->logger = $logger;
-
-        $this->converter = new Aggregate(
-            array(
-                new ToRichTextPreNormalize(new Expanding(), new EmbedLinking()),
-                new Xslt(
-                    './vendor/ezsystems/ezpublish-kernel/eZ/Publish/Core/FieldType/RichText/Resources/stylesheets/ezxml/docbook/docbook.xsl',
-                    array(
-                        array(
-                            'path' => './vendor/ezsystems/ezpublish-kernel/eZ/Publish/Core/FieldType/RichText/Resources/stylesheets/ezxml/docbook/core.xsl',
-                            'priority' => 99,
-                        ),
-                    )
-                ),
-            )
-        );
-
-        $this->validator = new Validator(
-            array(
-                './vendor/ezsystems/ezpublish-kernel/eZ/Publish/Core/FieldType/RichText/Resources/schemas/docbook/ezpublish.rng',
-                './vendor/ezsystems/ezpublish-kernel/eZ/Publish/Core/FieldType/RichText/Resources/schemas/docbook/docbook.iso.sch.xsl',
-            )
-        );
     }
 
     protected function configure()
@@ -182,6 +143,7 @@ EOT
 
     function convertFields($dryRun, $contentObjectId, $checkDuplicateIds, OutputInterface $output)
     {
+        $converter = new RichTextConverter($this->logger);
         $query = $this->db->createSelectQuery();
         $query->select($query->expr->count('*'));
         $query->from('ezcontentobject_attribute');
@@ -244,7 +206,7 @@ EOT
                 $inputValue = $row['data_text'];
             }
 
-            $converted = $this->convert($inputValue, $checkDuplicateIds, $row['id']);
+            $converted = $converter->convert($this->createDocument($inputValue), $checkDuplicateIds, $row['id']);
 
             $updateQuery = $this->db->createUpdateQuery();
             $updateQuery->update($this->db->quoteIdentifier('ezcontentobject_attribute'));
@@ -295,57 +257,5 @@ EOT
         $document->loadXml($xmlString);
 
         return $document;
-    }
-
-    function removeComments(DOMDocument $document)
-    {
-        $xpath = new DOMXpath($document);
-        $nodes = $xpath->query('//comment()');
-
-        for ($i = 0; $i < $nodes->length; ++$i) {
-            $nodes->item($i)->parentNode->removeChild($nodes->item($i));
-        }
-    }
-
-    function reportNonUniqueIds(DOMDocument $document, $contentObjectAttributeId)
-    {
-        $xpath = new DOMXPath($document);
-        $ns = $document->documentElement->namespaceURI;
-        $nodes = $xpath->query("//*[contains(@xml:id, 'duplicated_id_')]");
-        foreach ($nodes as $node) {
-            $id=$node->attributes->getNamedItem('id')->nodeValue;
-            // id has format "duplicated_id_foo_bar_idm45226413447104" where "foo_bar" is the duplicated id
-            $duplicatedId = substr($id, strlen('duplicated_id_'), strrpos($id, '_') - strlen('duplicated_id_'));
-            $this->logger->warning("Duplicated id in original ezxmltext for contentobject_attribute.id=$contentObjectAttributeId, automatically generated new id : $duplicatedId --> $id");
-        }
-    }
-
-    function convert($xmlString, $checkDuplicateIds, $contentObjectAttributeId)
-    {
-        $inputDocument = $this->createDocument($xmlString);
-
-        $this->removeComments($inputDocument);
-
-        $convertedDocument = $this->converter->convert($inputDocument);
-        if ($checkDuplicateIds) {
-            $this->reportNonUniqueIds($convertedDocument, $contentObjectAttributeId);
-        }
-
-        // Needed by some disabled output escaping (eg. legacy ezxml paragraph <line/> elements)
-        $convertedDocumentNormalized = new DOMDocument();
-        $convertedDocumentNormalized->loadXML($convertedDocument->saveXML());
-
-        $errors = $this->validator->validate($convertedDocument);
-
-        $result = $convertedDocumentNormalized->saveXML();
-
-        if (!empty($errors)) {
-            $this->logger->error(
-                "Validation errors when converting xmlstring",
-                ['result' => $result, 'errors' => $errors, 'xmlString' => $xmlString]
-            );
-        }
-
-        return $result;
     }
 }
